pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Simone-Alghisi <alghisius.simone@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
python-telegram-bot = "^13.2"
python-dotenv = "^0.15.0"
pandas = "1.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
